// components/layout/src/gqlTypes.d.ts
export type JSON = any;
// ...
export enum EmployeeType {
  AD = 'AD',
  RE = 'RE',
  HC = 'HC',
  IN = 'IN',
  SH = 'SH',
  TE = 'TE',
}
// ...
export namespace ClientDrawerContentQuery {
  export type Variables = {};

  export type Query = {
    __typename?: 'Query';
    dashboard: Dashboard;
  };

  export type Dashboard = {
    __typename?: 'Dashboard';
    id?: string | null;
    isSpoofing?: boolean | null;
    isTrialCompany?: boolean | null;
    switches: JSON;
    permission?: JSON | null;
    employee?: Employee | null;
    zAppInstallSubscriptions?: (ZAppInstallSubscriptions | null)[] | null;
  };

  export type Employee = {
    __typename?: 'AllEmployee';
    id?: string | null;
    type?: EmployeeType | null;
  };

  export type ZAppInstallSubscriptions = {
    __typename?: 'ZAppInstallSubscription';
    status?: number | null;
    inheritedStatus?: number | null;
    preferences?: JSON | null;
    appInstall?: AppInstall | null;
  };

  export type AppInstall = {
    __typename?: 'ZAppInstall';
    status?: number | null;
    preferences?: JSON | null;
    app?: App | null;
  };

  export type App = {
    __typename?: 'ZApp';
    uniqueId?: string | null;
    appUrl?: string | null;
    role?: string | null;
    appIconSqUrl?: string | null;
    shortTitle?: string | null;
    status?: number | null;
    preferences?: JSON | null;
  };
}